fixed unhandled par to int for default type: byte

+++ src/LoRaPhy/LoRaTransmitter.cc
@@ -45,7 +45,7 @@ void LoRaTransmitter::initialize(int stage)
         {
             iAmGateway = true;
         } else iAmGateway = false;
-        payloaddatasize = par('payloaddatasize');
+        payloaddatasize = par("payloaddatasize").intValue();
     }
 }
